Attempt to move to create-react-context
