+++ .gitignore
@@ -6,3 +6,4 @@ coverage
 .DS_Store
 next.d.ts
 legacy.d.ts
+.rpt2_cache
