+++ package.json
@@ -34,10 +34,11 @@
     "gen-docs": "all-contributors generate && doctoc README.md"
   },
   "dependencies": {
+    "create-react-context": "^0.1.6",
     "lodash.clonedeep": "^4.5.0",
     "lodash.isequal": "4.5.0",
     "lodash.topath": "4.5.2",
-    "prop-types": "^15.5.10",
+    "prop-types": "^15.6.0",
     "warning": "^3.0.0"
   },
   "peerDependencies": {
@@ -51,7 +52,6 @@
     "@types/lodash.clonedeep": "^4.5.3",
     "@types/lodash.isequal": "4.5.2",
     "@types/lodash.topath": "4.5.3",
-    "@types/prop-types": "15.5.1",
     "@types/react": "16.0.28",
     "@types/react-dom": "^16.0.3",
     "@types/react-test-renderer": "15.5.2",
