+++ src/index.tsx
@@ -1,0 +1,7 @@
+export * from './Formik';
+export * from './connect';
+export * from './Field';
+export * from './Form';
+export * from './withFormik';
+export * from './FieldArray';
+export * from './utils';
